chore(BUX-296): refactor code
